Removi global.c e .h
